MNT move common metadata routing test objects
